requirements.txt
# API Server Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Container Management
kubernetes==29.0.0
docker==7.1.0
requests==2.32.3
urllib3==2.2.2

# Background Tasks
celery==5.3.6
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1

# Logging
python-json-logger==2.0.7
